dev-requirements.txt
#
# Development dependencies.
#
nose==1.3.4
mock==1.0.1
coveralls==0.5
coverage==3.7.1
sphinx==1.4.8
sphinx-rtd-theme==0.1.9
python-dotenv==0.20.0